// src/types/generated/MetricsSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DiskUsage } from "./DiskUsage";

export type MetricsSnapshot = { serverId: string, timestamp: number, cpuPercent: number, cpuCores: number, memTotal: number, memUsed: number, memAvailable: number, memCached: number, swapTotal: number, swapUsed: number, load1: number, load5: number, load15: number, uptimeSecs: number, netRxBps: number, netTxBps: number, netRxTotal: number, netTxTotal: number, disks: Array<DiskUsage>, processCount: number, listeningPorts: number | null, latencyMs: number | null, };